test:fp16

+++ compare.cu
@@ -31,8 +31,9 @@
 // Therefore EPSILON = 0.0f is OK
 #define EPSILON 0.001f
 #define EPSILOND 0.0000001
+#include "cuda_fp16.h"
 
-extern "C" __global__ void compare(float *C, int *faultyElems, size_t iters) {
+extern "C" __global__ void compare(__half *C, int *faultyElems, size_t iters) {
 	size_t iterStep = blockDim.x*blockDim.y*gridDim.x*gridDim.y;
 	size_t myIndex = (blockIdx.y*blockDim.y + threadIdx.y)* // Y
 		gridDim.x*blockDim.x + // W
